add updateBunkerRot to reducer
